% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{metricsThreshold}
\alias{metricsThreshold}
\title{Computes some performance metrics for estimate of connectiviy matrix A.}
\usage{
metricsThreshold(trueA, est, thres = seq(0.01, 1, by = 0.01))
}
\arguments{
\item{trueA}{}

\item{est}{}

\item{thres}{}
}
\value{
res
}
\description{
Computes some performance metrics for estimate of connectiviy matrix A.
}

