% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotInterventionVar.R
\name{plotInterventionVars}
\alias{plotInterventionVars}
\title{Plots the estimated intervention variances.}
\usage{
plotInterventionVars(estIntVars, trueIntVars = NULL)
}
\arguments{
\item{estIntVars}{A  (Gxp)-dimensional matrix with the estimated intervention
variances returned by \code{backShift} (as \code{varianceEnv}).
\code{G} is the number of unique environments, \code{p} is the number of variables.}

\item{trueIntVars}{A (Gxp)-dimensional matrix with the true intervention variances
 if these are known (for simulations). By default this paramter is set to NULL.}
}
\description{
Plots the estimated intervention variances.
}

