% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateInterventions}
\alias{simulateInterventions}
\title{Simulate data of a causal cyclic model under shift interventions.}
\usage{
simulateInterventions(n, p, A, G, intervMultiplier, noiseMult, nonGauss,
  fracVarInt, hiddenVars, knownInterventions, simulateObs, seed = 1)
}
\arguments{
\item{n}{Number of observations.}

\item{p}{Number of variables.}

\item{A}{Connectivity matrix A. The entry \eqn{A_{ij}} contains the edge
from node i to node j.}

\item{G}{Number of environments.}

\item{intervMultiplier}{Regulates the strength of the interventions.}

\item{noiseMult}{Regulates the noise variance.}

\item{nonGauss}{Set to \code{TRUE} to generate non-Gaussian noise.}

\item{fracVarInt}{}

\item{hiddenVars}{}

\item{knownInterventions}{}

\item{simulateObs}{}

\item{seed}{}
}
\value{
A list with the following elements:
\itemize{
  \item X
  \item trueA
  \item environment
  \item interventions
  \item G
  \item indexObservationalData
  \item varInts
  \item intervMultiplier
  \item noiseMult
  \item fracVarInt
  \item hiddenVars
  \item knownInterventions
  \item simulateObs
}
}
\description{
Learning causal cyclic graphs from unknown shift interventions
}
\details{
Learning causal cyclic graphs from unknown shift interventions
}
\examples{
add(1, 1)
add(10, 1)
}

