% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotGraph.R
\name{plotGraphEdgeAttr}
\alias{plotGraphEdgeAttr}
\title{Plots graph from adjacency matrix.}
\usage{
plotGraphEdgeAttr(pointEst, labels, thres.point, adjacencyEst, plotStabSelec,
  thres.stab, main, ...)
}
\arguments{
\item{pointEst}{}

\item{labels}{}

\item{thres.point}{}

\item{adjacencyEst}{}

\item{plotStabSelec}{}

\item{thres.stab}{}

\item{main}{}

\item{...}{}
}
\description{
Plots graph from adjacency matrix.
}

